requirements.txt
flask
python-dotenv
flask-smorest
marshmallow
mysql.connector
flask-jwt-extended
